:recycle: refactor: 프로젝트명 변경
Issue: #14
