+++ app/src/main/java/com/bibbidi/habittracking/ui/customview/DayOfTheWeek.kt
@@ -1,4 +1,4 @@
-package com.bibbidi.myrootineclone.ui.customview
+package com.bibbidi.habittracking.ui.customview
 
 enum class DayOfTheWeek {
     SUN,
