+++ app/src/main/res/values-night/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.MyRootineClone" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.HabitTracking" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/green_03</item>
         <item name="colorPrimaryVariant">@color/green_02</item>
