+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.MyRootineClone" parent="Theme.MaterialComponents.Light.NoActionBar">
+    <style name="Theme.HabitTracking" parent="Theme.MaterialComponents.Light.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/green_03</item>
         <item name="colorPrimaryVariant">@color/green_02</item>
